refactor: error 클래스 파일 분리 & 중복 파일 삭제
