+++ src/common/errors/app-error.ts
@@ -1,0 +1,11 @@
+// 커스텀 에러 클래스
+export class AppError extends Error {
+  constructor(
+    public message: string,
+    public statusCode: number,
+  ) {
+    super(message);
+    this.name = this.constructor.name;
+    Error.captureStackTrace(this, this.constructor);
+  }
+}
