+++ src/common/errors/bad-request-error.ts
@@ -1,0 +1,7 @@
+import { AppError } from './app-error';
+
+export class BadRequestError extends AppError {
+  constructor(message: string) {
+    super(message, 400);
+  }
+}
