+++ src/common/errors/conflict-error.ts
@@ -1,0 +1,7 @@
+import { AppError } from './app-error';
+
+export class ConflictError extends AppError {
+  constructor(message: string) {
+    super(message, 409);
+  }
+}
