+++ src/common/errors/forbidden-error.ts
@@ -1,0 +1,7 @@
+import { AppError } from './app-error';
+
+export class ForbiddenError extends AppError {
+  constructor(message: string) {
+    super(message, 403);
+  }
+}
