+++ src/common/errors/unauthorize-error.ts
@@ -1,0 +1,7 @@
+import { AppError } from './app-error';
+
+export class UnauthorizedError extends AppError {
+  constructor(message: string) {
+    super(message, 401);
+  }
+}
