--- src/common/utils/custom-errors.ts
@@ -1,35 +1,0 @@
-// eslint-disable-next-line max-classes-per-file
-export class AppError extends Error {
-  constructor(
-    public message: string,
-    public statusCode: number,
-  ) {
-    super(message);
-    this.name = this.constructor.name;
-    Error.captureStackTrace(this, this.constructor);
-  }
-}
-
-export class NotFoundError extends AppError {
-  constructor(message: string) {
-    super(message, 404);
-  }
-}
-
-export class BadRequestError extends AppError {
-  constructor(message: string) {
-    super(message, 400);
-  }
-}
-
-export class ConflictError extends AppError {
-  constructor(message: string) {
-    super(message, 409);
-  }
-}
-
-export class UnauthorizedError extends AppError {
-  constructor(message: string) {
-    super(message, 401);
-  }
-}
