+++ src/middlewares/error.middleware.ts
@@ -1,50 +1,9 @@
 /* eslint-disable */
 import { Request, Response, NextFunction } from 'express';
+import { AppError } from '../common/errors/app-error';
 
 interface CustomError extends Error {
   status?: number;
-}
-
-// 커스텀 에러 클래스
-export class AppError extends Error {
-  constructor(
-    public message: string,
-    public statusCode: number,
-  ) {
-    super(message);
-    this.name = this.constructor.name;
-    Error.captureStackTrace(this, this.constructor);
-  }
-}
-
-export class NotFoundError extends AppError {
-  constructor(message: string) {
-    super(message, 404);
-  }
-}
-
-export class BadRequestError extends AppError {
-  constructor(message: string) {
-    super(message, 400);
-  }
-}
-
-export class ConflictError extends AppError {
-  constructor(message: string) {
-    super(message, 409);
-  }
-}
-
-export class UnauthorizedError extends AppError {
-  constructor(message: string) {
-    super(message, 401);
-  }
-}
-
-export class ForbiddenError extends AppError {
-  constructor(message: string) {
-    super(message, 403);
-  }
 }
 
 // 통합 에러 핸들러
